refactor: rename file travel prompt
