+++ src/mcp_server/components/prompts/travel_prompts.py
@@ -1,5 +1,5 @@
 from mcp_server.mcp_instance import mcp
-from mcp_server.core.prompts.get_travel_prompts import get_itinerary_prompt
+from mcp_server.core.prompts.travel import get_itinerary_prompt
 
 @mcp.prompt()
 async def generate_itinerary_prompt(days: int, start_date: str) -> str:
