+++ src/mcp_server/core/prompts/travel.py
